+++ handlers/index.go
@@ -3,7 +3,7 @@ package handlers
 import (
 	"net/http"
 
-	"github.com/daison12006013/gorvel/glade"
+	"github.com/daison12006013/gorvel/internals/view"
 	"github.com/daison12006013/gorvel/structs"
 	"github.com/gorilla/mux"
 )
@@ -16,5 +16,5 @@ func Home(w http.ResponseWriter, r *http.Request) {
 		Params: mux.Vars(r),
 	}
 
-	glade.View(w, "views/index.html", data)
+	view.Render(w, "views/index.html", data)
 }
