+++ internals/view/view.go
@@ -1,11 +1,11 @@
-package glade
+package view
 
 import (
 	"net/http"
 	"text/template"
 )
 
-func View(w http.ResponseWriter, file string, data interface{}) {
+func Render(w http.ResponseWriter, file string, data interface{}) {
 	t, _ := template.ParseFiles(file)
 	t.Execute(w, data)
 }
